+++ CHANGELOG.md
@@ -41,3 +41,4 @@ Fix the bug in the app installer where it fails after unzipping the app file.
 - Added a check for venv in the installer.
 - fixed bug in installer where it didn't install the app correctly after unzipping the app file. (cause: wrong path for pyapp.toml)
 - fixed bug where "added repository" message was shown when loading the repository file.
+- fixed bug where pyappm.toml was left dangling in /tmp/pyappm after installation.